Introduserer en ny REST tjeneste med nye Dtoer
